+++ elixir-generate-dockerfile/src/test/generator_test.exs
@@ -35,13 +35,13 @@ defmodule GeneratorTest do
     assert_dockerfile_line("## Project: (unknown)")
     assert_dockerfile_line("FROM gcr.io/gcp-elixir/runtime/ubuntu18/builder AS app-build")
     assert_dockerfile_line("#     && apt-get install -y -q package-name")
-    assert_dockerfile_line("ARG erlang_version=\"20.2\"")
-    assert_dockerfile_line("ARG elixir_version=\"1.5.3-otp-20\"")
+    assert_dockerfile_line("ARG erlang_version=\"22.0.7\"")
+    assert_dockerfile_line("ARG elixir_version=\"1.9.0-otp-22\"")
     assert_dockerfile_line("RUN asdf plugin-update erlang")
     assert_dockerfile_line("# RUN gcloud config set project my-project-id")
     assert_dockerfile_line("# ENV NAME=\"value\"")
     assert_dockerfile_line("# ENV BUILD_CLOUDSQL_INSTANCES=\"my-project-id:db-region:db-name\"")
-    refute_dockerfile_line("RUN mix release --env=prod --verbose")
+    refute_dockerfile_line("RUN mix distillery.release --env=prod --verbose")
     assert_dockerfile_line("FROM gcr.io/gcp-elixir/runtime/ubuntu18/asdf")
     assert_dockerfile_line("CMD exec mix run --no-halt")
   end
@@ -149,13 +149,13 @@ defmodule GeneratorTest do
 
     run_generator("minimal", config)
     assert_dockerfile_line("FROM gcr.io/gcp-elixir/runtime/ubuntu18/builder AS app-build")
-    assert_dockerfile_line("ARG erlang_version=\"20.2\"")
-    assert_dockerfile_line("ARG elixir_version=\"1.5.3-otp-20\"")
+    assert_dockerfile_line("ARG erlang_version=\"22.0.7\"")
+    assert_dockerfile_line("ARG elixir_version=\"1.9.0-otp-22\"")
     assert_dockerfile_line("RUN asdf plugin-update erlang")
-    assert_dockerfile_line("RUN mix release --env=prod --verbose")
+    assert_dockerfile_line("RUN mix release")
     assert_dockerfile_line("COPY --from=app-build /app/_build/prod/rel/my_app /app/")
     assert_dockerfile_line("FROM gcr.io/gcp-elixir/runtime/ubuntu18")
-    assert_dockerfile_line("CMD [\"/app/bin/my_app\",\"foreground\"]")
+    assert_dockerfile_line("CMD [\"/app/bin/my_app\",\"start\"]")
   end
 
   test "minimal directory with release app and custom entrypoint" do
@@ -164,12 +164,12 @@ defmodule GeneratorTest do
         """
         runtime_config:
           release_app: my_app
-        entrypoint: /app/bin/my_app foreground --blah
+        entrypoint: /app/bin/my_app start --blah
         """
 
     run_generator("minimal", config)
-    assert_dockerfile_line("RUN mix release --env=prod --verbose")
-    assert_dockerfile_line("CMD exec /app/bin/my_app foreground --blah")
+    assert_dockerfile_line("RUN mix release my_app")
+    assert_dockerfile_line("CMD exec /app/bin/my_app start --blah")
   end
 
   test "minimal directory with release app and custom mix_env" do
@@ -183,7 +183,7 @@ defmodule GeneratorTest do
         """
 
     run_generator("minimal", config)
-    assert_dockerfile_line("RUN mix release --env=staging --verbose")
+    assert_dockerfile_line("RUN mix release my_app")
     assert_dockerfile_line("COPY --from=app-build /app/_build/staging/rel/my_app /app/")
   end
 
@@ -191,25 +191,28 @@ defmodule GeneratorTest do
     run_generator("phoenix_1_2", @minimal_config)
     assert_ignore_line("priv/static")
     assert_ignore_line("node_modules")
-    refute_dockerfile_line("RUN mix release --env=prod --verbose")
+    refute_dockerfile_line("RUN mix release")
+    refute_dockerfile_line("RUN mix distillery.release")
   end
 
   test "phoenix 1.3 directory with custom elixir" do
     run_generator("phoenix_1_3", @minimal_config)
     assert_ignore_line("priv/static")
     assert_ignore_line("assets/node_modules")
-    assert_dockerfile_line("ARG erlang_version=\"20.2\"")
-    assert_dockerfile_line("ARG elixir_version=\"1.5.1\"")
-    refute_dockerfile_line("RUN mix release --env=prod --verbose")
+    assert_dockerfile_line("ARG erlang_version=\"21.1\"")
+    assert_dockerfile_line("ARG elixir_version=\"1.7.4-otp-21\"")
+    refute_dockerfile_line("RUN mix release")
+    refute_dockerfile_line("RUN mix distillery.release")
   end
 
   test "phoenix umbrella 1.3 directory with custom erlang and elixir" do
     run_generator("phoenix_umbrella_1_3", @minimal_config)
     assert_ignore_line("priv/static")
     assert_ignore_line("apps/blog_web/assets/node_modules")
-    assert_dockerfile_line("ARG erlang_version=\"20.0\"")
-    assert_dockerfile_line("ARG elixir_version=\"1.5.1-otp-20\"")
-    refute_dockerfile_line("RUN mix release --env=prod --verbose")
+    assert_dockerfile_line("ARG erlang_version=\"21.1\"")
+    assert_dockerfile_line("ARG elixir_version=\"1.8.2-otp-21\"")
+    refute_dockerfile_line("RUN mix release")
+    refute_dockerfile_line("RUN mix distillery.release")
   end
 
   test "phoenix 1.3 directory with release app and custom elixir" do
@@ -221,8 +224,8 @@ defmodule GeneratorTest do
         """
 
     run_generator("phoenix_1_3", config)
-    assert_dockerfile_line("ARG erlang_version=\"20.2\"")
-    assert_dockerfile_line("ARG elixir_version=\"1.5.1\"")
+    assert_dockerfile_line("ARG erlang_version=\"21.1\"")
+    assert_dockerfile_line("ARG elixir_version=\"1.7.4-otp-21\"")
     assert_dockerfile_line("RUN mix release --env=prod --verbose")
   end
 
@@ -235,18 +238,48 @@ defmodule GeneratorTest do
         """
 
     run_generator("phoenix_umbrella_1_3", config)
-    assert_dockerfile_line("ARG erlang_version=\"20.0\"")
-    assert_dockerfile_line("ARG elixir_version=\"1.5.1-otp-20\"")
-    assert_dockerfile_line("RUN mix release --env=prod --verbose")
+    assert_dockerfile_line("ARG erlang_version=\"21.1\"")
+    assert_dockerfile_line("ARG elixir_version=\"1.8.2-otp-21\"")
+    assert_dockerfile_line("RUN mix release blog")
+  end
+
+  test "phoenix 1.3 directory with release app and custom mix_env" do
+    config =
+      @minimal_config <>
+        """
+        runtime_config:
+          release_app: blog
+        env_variables:
+          MIX_ENV: staging
+        """
+
+    run_generator("phoenix_1_3", config)
+    assert_dockerfile_line("RUN mix release --env=staging --verbose")
+    assert_dockerfile_line("COPY --from=app-build /app/_build/staging/rel/blog /app/")
+  end
+
+  test "phoenix 1.4 directory with release app and custom mix_env" do
+    config =
+      @minimal_config <>
+        """
+        runtime_config:
+          release_app: blog
+        env_variables:
+          MIX_ENV: staging
+        """
+
+    run_generator("phoenix_1_4", config)
+    assert_dockerfile_line("RUN mix distillery.release --env=staging --verbose")
+    assert_dockerfile_line("COPY --from=app-build /app/_build/staging/rel/blog /app/")
   end
 
   test "minimal directory with prebuilt erlang" do
-    run_generator("minimal", @minimal_config, prebuilt_erlang_versions: "20.2")
-    assert_dockerfile_line("ARG erlang_version=\"20.2\"")
-    assert_dockerfile_line("ARG elixir_version=\"1.5.3-otp-20\"")
+    run_generator("minimal", @minimal_config, prebuilt_erlang_versions: "22.0.7")
+    assert_dockerfile_line("ARG erlang_version=\"22.0.7\"")
+    assert_dockerfile_line("ARG elixir_version=\"1.9.0-otp-22\"")
 
     assert_dockerfile_line(
-      "COPY --from=gcr.io/gcp-elixir/runtime/ubuntu18/prebuilt/otp-20.2:latest"
+      "COPY --from=gcr.io/gcp-elixir/runtime/ubuntu18/prebuilt/otp-22.0.7:latest"
     )
   end
 
@@ -258,12 +291,12 @@ defmodule GeneratorTest do
           release_app: my_app
         """
 
-    run_generator("minimal", config, prebuilt_erlang_versions: "20.2")
-    assert_dockerfile_line("ARG erlang_version=\"20.2\"")
-    assert_dockerfile_line("ARG elixir_version=\"1.5.3-otp-20\"")
+    run_generator("minimal", config, prebuilt_erlang_versions: "22.0.7")
+    assert_dockerfile_line("ARG erlang_version=\"22.0.7\"")
+    assert_dockerfile_line("ARG elixir_version=\"1.9.0-otp-22\"")
 
     assert_dockerfile_line(
-      "COPY --from=gcr.io/gcp-elixir/runtime/ubuntu18/prebuilt/otp-20.2:latest"
+      "COPY --from=gcr.io/gcp-elixir/runtime/ubuntu18/prebuilt/otp-22.0.7:latest"
     )
   end
 
@@ -305,8 +338,8 @@ defmodule GeneratorTest do
       os_image: "gcr.io/gcp-elixir/runtime/ubuntu18",
       asdf_image: "gcr.io/gcp-elixir/runtime/ubuntu18/asdf",
       builder_image: "gcr.io/gcp-elixir/runtime/ubuntu18/builder",
-      default_erlang_version: "20.2",
-      default_elixir_version: "1.5.3-otp-20"
+      default_erlang_version: "22.0.7",
+      default_elixir_version: "1.9.0-otp-22"
     ]
 
     opts =