+++ test/sample_apps/minimal_phoenix14/mix-dist20.exs
@@ -3,10 +3,9 @@ defmodule MinimalPhoenix14.MixProject do
 
   def project do
     [
-      releases: releases(),
       app: :minimal_phoenix14,
       version: "0.1.0",
-      elixir: "~> 1.9",
+      elixir: "~> 1.8",
       elixirc_paths: elixirc_paths(Mix.env()),
       compilers: [:phoenix, :gettext] ++ Mix.compilers(),
       start_permanent: Mix.env() == :prod,
@@ -32,17 +31,8 @@ defmodule MinimalPhoenix14.MixProject do
       {:phoenix_live_reload, "~> 1.2", only: :dev},
       {:gettext, "~> 0.11"},
       {:jason, "~> 1.0"},
-      {:plug_cowboy, "~> 2.0"}
-    ]
-  end
-  
-  defp releases do
-    [
-      minimal_phoenix14: [
-        include_erts: true,
-        include_executables_for: [:unix],
-        applications: [runtime_tools: :permanent]
-      ]
+      {:plug_cowboy, "~> 2.0"},
+      {:distillery, "~> 2.0.0"}
     ]
   end
 end